Extract source map utils
