+++ packages/devtools-source-map/src/tests/getContentType.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const { getContentType, contentMapForTesting } = require("../util");
+const { getContentType, contentMapForTesting } = require("../utils");
 
 test("getContentType", () => {
   for (let ext in contentMapForTesting) {
