+++ packages/devtools-source-map/src/tests/isURL.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const { isURL } = require("../path");
+const { isURL } = require("../utils/path");
 
 test("isURL", () => {
   expect(isURL("http://www.example.com/")).toEqual(true);
