+++ packages/devtools-source-map/src/utils/path.js
